# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WiseEff API",
    "version": "m5"
  },
  "paths": {
...
    "/api/v1/log-files": {
      "post": {
        "operationId": "logs.uploadFile",
        "summary": "Upload log file",
        "tags": [
          "logs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogFileUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogFileUploadResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/logs": {
      "post": {
        "operationId": "logs.upload",
        "summary": "Create log analysis record",
        "tags": [
          "logs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLogRecordRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogRecordResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "get": {
        "operationId": "logs.list",
        "summary": "List log records",
        "tags": [
          "logs"
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogRecordListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/logs/{logId}": {
      "get": {
        "operationId": "logs.get",
        "summary": "Get log record",
        "tags": [
          "logs"
        ],
        "parameters": [
          {
            "name": "logId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogRecordResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/logs/{logId}/runs": {
      "get": {
        "operationId": "logs.listRuns",
        "summary": "List log analysis runs",
        "tags": [
          "logs"
        ],
        "parameters": [
          {
            "name": "logId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogRunListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/logs/{logId}/rerun": {
      "post": {
        "operationId": "logs.rerun",
        "summary": "Rerun log analysis",
        "tags": [
          "logs"
        ],
        "parameters": [
          {
            "name": "logId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogRunResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/logs/{logId}/archive": {
      "post": {
        "operationId": "logs.archive",
        "summary": "Archive log record",
        "tags": [
          "logs"
        ],
        "parameters": [
          {
            "name": "logId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogRecordResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/logs/{logId}/unarchive": {
      "post": {
        "operationId": "logs.unarchive",
        "summary": "Unarchive log record",
        "tags": [
          "logs"
        ],
        "parameters": [
          {
            "name": "logId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogRecordResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/logs/{logId}/feedback": {
      "post": {
        "operationId": "logs.feedback",
        "summary": "Submit log feedback",
        "tags": [
          "logs"
        ],
        "parameters": [
          {
            "name": "logId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogFeedbackResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "LogFileUploadRequest": {
        "type": "object",
        "x-wiseeff-schema": "LogFileUploadRequest"
      },
      "LogFileUploadResponse": {
        "type": "object",
        "x-wiseeff-schema": "LogFileUploadResponse"
      },
      "CreateLogRecordRequest": {
        "type": "object",
        "x-wiseeff-schema": "CreateLogRecordRequest"
      },
      "LogRecordResponse": {
        "type": "object",
        "x-wiseeff-schema": "LogRecordResponse"
      },
      "LogRecordListResponse": {
        "type": "object",
        "x-wiseeff-schema": "LogRecordListResponse"
      },
      "LogRunListResponse": {
        "type": "object",
        "x-wiseeff-schema": "LogRunListResponse"
      },
      "LogRunResponse": {
        "type": "object",
        "x-wiseeff-schema": "LogRunResponse"
      },
      "LogFeedbackRequest": {
        "type": "object",
        "x-wiseeff-schema": "LogFeedbackRequest"
      },
      "LogFeedbackResponse": {
        "type": "object",
        "x-wiseeff-schema": "LogFeedbackResponse"
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "WiseEff error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}